+++ src/components/Text/textChangelogData.ts
@@ -13,21 +13,15 @@ export const changelogData: ChangelogData[] = [
     date: "Prerelease",
     version: "Prerelease",
     type: "Update",
-    affects: ["Accessibility"],
+    affects: ["Styles"],
     notes: [
+      "Chakra 2.8 update.",
       "Updates the component's theme file to use native Chakra responsive styles for the font sizes of the `subtitle1` and `subtitle2` variants.",
     ],
   },
   {
     date: "2023-12-07",
     version: "2.1.3",
-    type: "Update",
-    affects: ["Styles"],
-    notes: ["Chakra 2.8 update."],
-  },
-  {
-    date: "Prerelease",
-    version: "Prerelease",
     type: "Update",
     affects: ["Accessibility"],
     notes: [