+++ src/theme/components/heading.ts
@@ -90,6 +90,10 @@ export const headings = {
   }),
   display1: definePartsStyle({
     base: {
+      fontSize: {
+        base: "mobile.heading.display1",
+        md: "desktop.heading.display1",
+      },
       fontWeight: "heading.display1",
       letterSpacing: "0.0625rem",
       lineHeight: "1.10",
@@ -105,6 +109,10 @@ export const headings = {
   }),
   heading1: definePartsStyle({
     base: {
+      fontSize: {
+        base: "mobile.heading.heading1",
+        md: "desktop.heading.heading1",
+      },
       fontWeight: "heading.heading1",
       letterSpacing: "0.0625rem",
       lineHeight: "1.15",
@@ -120,6 +128,10 @@ export const headings = {
   }),
   heading2: definePartsStyle({
     base: {
+      fontSize: {
+        base: "mobile.heading.heading2",
+        md: "desktop.heading.heading2",
+      },
       fontWeight: "heading.heading2",
       letterSpacing: "0.0625rem",
       lineHeight: "1.20",
@@ -135,6 +147,10 @@ export const headings = {
   }),
   heading3: definePartsStyle({
     base: {
+      fontSize: {
+        base: "mobile.heading.heading3",
+        md: "desktop.heading.heading3",
+      },
       fontWeight: "heading.heading3",
       letterSpacing: "0",
       lineHeight: "1.25",
@@ -150,6 +166,10 @@ export const headings = {
   }),
   heading4: definePartsStyle({
     base: {
+      fontSize: {
+        base: "mobile.heading.heading4",
+        md: "desktop.heading.heading4",
+      },
       fontWeight: "heading.heading4",
       letterSpacing: "0",
       lineHeight: "1.30",
@@ -165,6 +185,10 @@ export const headings = {
   }),
   heading5: definePartsStyle({
     base: {
+      fontSize: {
+        base: "mobile.heading.heading5",
+        md: "desktop.heading.heading5",
+      },
       fontWeight: "heading.heading5",
       letterSpacing: "0",
       lineHeight: "1.35",
@@ -179,6 +203,10 @@ export const headings = {
   }),
   heading6: definePartsStyle({
     base: {
+      fontSize: {
+        base: "mobile.heading.heading6",
+        md: "desktop.heading.heading6",
+      },
       fontWeight: "heading.heading6",
       letterSpacing: "0",
       lineHeight: "1.40",